+++ install_scripts/nvim.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 set -eux
 
+source "${DOTFILES_DIR:-$(cd "$(dirname "$0")/.." && pwd)}/lib/utils.sh"
+
 VERSION="0.10.4"
 PREFIX="${PWD}/neovim-${VERSION}/build"
-thread=15
 BIN=${BIN:-$HOME/works/bin}
+thread=$(detect_nproc)
+
+[ -d "neovim-${VERSION}" ] && { print_info "neovim-${VERSION} already present, skipping"; exit 0; }
 
 URL="https://github.com/neovim/neovim/archive/refs/tags/v${VERSION}.tar.gz"
 curl -L "$URL" -o "v${VERSION}.tar.gz"
@@ -14,6 +18,6 @@ cd neovim-${VERSION}
 make -j $thread CMAKE_BUILD_TYPE=RelWithDebInfo CMAKE_INSTALL_PREFIX=${PREFIX}
 make -j $thread install
 
-ln -s ${PREFIX}/bin/nvim $BIN
+ensure_bin ${PREFIX}/bin/nvim
 
-echo "[INFO] nvim install done" >&1
+print_info "nvim install done"
